refactor: 답변 작성자만 답변을 삭제할 수 있도록 수정

+++ src/main/java/qna/domain/Answer.java
@@ -1,6 +1,7 @@
 package qna.domain;
 
 import java.time.LocalDateTime;
+import qna.CannotDeleteException;
 import qna.NotFoundException;
 import qna.UnAuthorizedException;
 
@@ -69,7 +70,11 @@ public class Answer extends AbstractEntity {
         this.question = question;
     }
 
-    public DeleteHistory delete() {
+    public DeleteHistory deleteBy(User loginUser) {
+        if (!isOwner(loginUser)) {
+            throw new CannotDeleteException("답변을 삭제할 권한이 없습니다.");
+        }
+
         deleted = true;
         return new DeleteHistory(ContentType.ANSWER, getId(), writer, LocalDateTime.now());
     }
